HSC-1473: E2E Tests for services and routes checking on CF Dashboard (#959)

+++ e2e/po/helpers.po.js
@@ -155,8 +155,10 @@
     return browser.forkNewDriverInstance(true);
   }
 
-  function loadApp() {
-    browser.manage().deleteAllCookies();
+  function loadApp(keepCookies) {
+    if (!keepCookies) {
+      browser.manage().deleteAllCookies();
+    }
     return browser.get(host);
   }
